+++ activerecord/test/cases/tasks/database_tasks_test.rb
@@ -112,59 +112,6 @@ module ActiveRecord
     end
   end
 
-  class DatabaseTasksCurrentConfigTask < ActiveRecord::TestCase
-    def test_current_config_set
-      hash = {}
-
-      with_stubbed_configurations do
-        ActiveRecord::Tasks::DatabaseTasks.current_config(config: hash, env: "production")
-
-        assert_equal hash, ActiveRecord::Tasks::DatabaseTasks.current_config(env: "production")
-      end
-    end
-
-    def test_current_config_read_none_found
-      with_stubbed_configurations do
-        config = ActiveRecord::Tasks::DatabaseTasks.current_config(env: "production", spec: "empty")
-
-        assert_nil config
-      end
-    end
-
-    def test_current_config_read_found
-      with_stubbed_configurations do
-        config = ActiveRecord::Tasks::DatabaseTasks.current_config(env: "production", spec: "exists")
-
-        assert_equal({ database: "my-db" }, config)
-      end
-    end
-
-    def test_current_config_read_after_set
-      hash = {}
-
-      with_stubbed_configurations do
-        ActiveRecord::Tasks::DatabaseTasks.current_config(config: hash, env: "production")
-
-        config = ActiveRecord::Tasks::DatabaseTasks.current_config(env: "production", spec: "exists")
-
-        assert_equal hash, config
-      end
-    end
-
-    private
-      def with_stubbed_configurations(&block)
-        old_configurations = ActiveRecord::Base.configurations
-        ActiveRecord::Base.configurations = { "production" => { "exists" => { "database" => "my-db" } } }
-
-        assert_deprecated(&block)
-      ensure
-        ActiveRecord::Base.configurations = old_configurations
-        assert_deprecated do
-          ActiveRecord::Tasks::DatabaseTasks.current_config = nil
-        end
-      end
-  end
-
   class DatabaseTasksRegisterTask < ActiveRecord::TestCase
     def test_register_task
       klazz = Class.new do
@@ -1597,38 +1544,10 @@ module ActiveRecord
       @configurations = { "development" => { "database" => "my-db" } }
     end
 
-    def test_check_schema_file_defaults
-      ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "/tmp") do
-        assert_deprecated do
-          assert_equal "/tmp/schema.rb", ActiveRecord::Tasks::DatabaseTasks.schema_file
-        end
-      end
-    end
-
-    { ruby: "schema.rb", sql: "structure.sql" }.each_pair do |fmt, filename|
-      define_method("test_check_schema_file_for_#{fmt}_format") do
-        ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "/tmp") do
-          assert_deprecated do
-            assert_equal "/tmp/#{filename}", ActiveRecord::Tasks::DatabaseTasks.schema_file(fmt)
-          end
-        end
-      end
-    end
-
     def test_check_dump_filename_defaults
       ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "/tmp") do
         with_stubbed_configurations do
           assert_equal "/tmp/schema.rb", ActiveRecord::Tasks::DatabaseTasks.schema_dump_path(config_for("development", "primary"))
-        end
-      end
-    end
-
-    def test_dump_filename_is_deprecated
-      ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "/tmp") do
-        with_stubbed_configurations do
-          assert_deprecated do
-            assert_equal "/tmp/schema.rb", ActiveRecord::Tasks::DatabaseTasks.dump_filename(config_for("development", "primary").name)
-          end
         end
       end
     end